Initial work on configurable api-root
Now it works with 'cised-ca/social-enterprise-directory', but breaks the
local 'src/data/' mock API.

I believe this is because webpack-dev-server (or our current
configuration of it) doesn't like serving extension-less assets.

I can't seem to find a way to alias or rewrite paths either.

We don't _really_ need the mock API since we can run the backend locally
too, but it would be kind of nice to have this somewhat stand-alone.

Ref. #12